Remove ByteBuffer[] related methods from ChannelOutboundBuffer
Motivation:

The ByteBuffer[] related methods are only needed for NioSocketChannel and can be implemented in a MessageProcessor. Let's remove these from ChannelOutboundBuffer.

Modifications:

Move ByteBuffer[] collecting into a MessageProcessor implementation that is hidden by the end-user.

Result:

Cleanup of API
