+++ transport/src/test/java/io/netty5/channel/ChannelOutboundBufferTest.java
@@ -17,7 +17,6 @@ package io.netty5.channel;
 
 import io.netty5.buffer.api.Buffer;
 import io.netty5.buffer.api.BufferAllocator;
-import io.netty5.buffer.api.CompositeBuffer;
 import io.netty5.channel.ChannelOutboundBuffer.MessageProcessor;
 import io.netty5.util.CharsetUtil;
 import io.netty5.util.concurrent.EventExecutor;
@@ -25,11 +24,8 @@ import io.netty5.util.concurrent.Promise;
 import io.netty5.util.concurrent.SingleThreadEventExecutor;
 import org.junit.jupiter.api.Test;
 
-import java.nio.ByteBuffer;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.BiConsumer;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -58,19 +54,6 @@ public class ChannelOutboundBufferTest {
     }
 
     @Test
-    public void testEmptyNioBuffers() throws InterruptedException {
-        testChannelOutboundBuffer((buffer, executor) -> {
-            assertEquals(0, buffer.nioBufferCount());
-            ByteBuffer[] buffers = buffer.nioBuffers();
-            assertNotNull(buffers);
-            for (ByteBuffer b : buffers) {
-                assertNull(b);
-            }
-            assertEquals(0, buffer.nioBufferCount());
-        });
-    }
-
-    @Test
     public void flushingEmptyBuffers() throws InterruptedException {
         testChannelOutboundBuffer((buffer, executor) -> {
             Buffer buf = BufferAllocator.onHeapUnpooled().allocate(0);
@@ -88,114 +71,6 @@ public class ChannelOutboundBufferTest {
             messageCounter.set(0);
             buffer.forEachFlushedMessage(messageProcessor);
             assertThat(messageCounter.get()).isZero();
-        });
-    }
-
-    @Test
-    public void testNioBuffersSingleBacked() throws InterruptedException {
-        testChannelOutboundBuffer((buffer, executor) -> {
-            assertEquals(0, buffer.nioBufferCount());
-
-            Buffer buf = BufferAllocator.onHeapUnpooled().copyOf("buf1", CharsetUtil.US_ASCII);
-            buffer.addMessage(buf, buf.readableBytes(), executor.newPromise());
-            assertEquals(0, buffer.nioBufferCount(), "Should still be 0 as not flushed yet");
-            buffer.addFlush();
-            ByteBuffer[] buffers = buffer.nioBuffers();
-            assertNotNull(buffers);
-            assertEquals(1, buffer.nioBufferCount(), "Should still be 0 as not flushed yet");
-            for (int i = 0; i < buffer.nioBufferCount(); i++) {
-                if (i == 0) {
-                    assertEquals(1, buf.countReadableComponents());
-                    try (var iteration = buf.forEachComponent()) {
-                        var component = iteration.firstReadable();
-                        assertEquals(buffers[0], component.readableBuffer());
-                        assertNull(component.nextReadable(), "Expected buffer to only have a single component.");
-                    }
-                } else {
-                    assertNull(buffers[i]);
-                }
-            }
-        });
-    }
-
-    @Test
-    public void testNioBuffersExpand() throws InterruptedException {
-        testChannelOutboundBuffer((buffer, executor) -> {
-            Buffer buf = BufferAllocator.offHeapUnpooled().copyOf("buf1", CharsetUtil.US_ASCII);
-            for (int i = 0; i < 64; i++) {
-                buffer.addMessage(buf.copy(), buf.readableBytes(), executor.newPromise());
-            }
-            assertEquals(0, buffer.nioBufferCount(), "Should still be 0 as not flushed yet");
-            buffer.addFlush();
-            ByteBuffer[] buffers = buffer.nioBuffers();
-            assertEquals(64, buffer.nioBufferCount());
-            assertEquals(1, buf.countReadableComponents());
-            try (var iteration = buf.forEachComponent()) {
-                var component = iteration.firstReadable();
-                ByteBuffer expected = component.readableBuffer();
-                for (int i = 0; i < buffer.nioBufferCount(); i++) {
-                    assertEquals(expected, buffers[i]);
-                }
-                assertNull(component.nextReadable());
-            }
-            buf.close();
-        });
-    }
-
-    @Test
-    public void testNioBuffersExpand2() throws InterruptedException {
-        testChannelOutboundBuffer((buffer, executor) -> {
-            Buffer buf = BufferAllocator.offHeapUnpooled().copyOf("buf1", CharsetUtil.US_ASCII);
-            var sends = Stream.generate(() -> buf.copy().send()).limit(65).collect(Collectors.toList());
-            CompositeBuffer comp = BufferAllocator.offHeapUnpooled().compose(sends);
-
-            buffer.addMessage(comp, comp.readableBytes(), executor.newPromise());
-
-            assertEquals(0, buffer.nioBufferCount(), "Should still be 0 as not flushed yet");
-            buffer.addFlush();
-            ByteBuffer[] buffers = buffer.nioBuffers();
-            assertEquals(65, buffer.nioBufferCount());
-            assertEquals(1, buf.countReadableComponents());
-            try (var iteration = buf.forEachComponent()) {
-                var component = iteration.firstReadable();
-                ByteBuffer expected = component.readableBuffer();
-                for (int i = 0; i < buffer.nioBufferCount(); i++) {
-                    if (i < 65) {
-                        assertEquals(expected, buffers[i]);
-                    } else {
-                        assertNull(buffers[i]);
-                    }
-                }
-                assertNull(component.nextReadable());
-            }
-            buf.close();
-        });
-    }
-
-    @Test
-    public void testNioBuffersMaxCount() throws InterruptedException {
-        testChannelOutboundBuffer((buffer, executor) -> {
-            Buffer buf = BufferAllocator.offHeapUnpooled().copyOf("buf1", CharsetUtil.US_ASCII);
-            assertEquals(4, buf.readableBytes());
-            var sends = Stream.generate(() -> buf.copy().send()).limit(65).collect(Collectors.toList());
-            CompositeBuffer comp = BufferAllocator.offHeapUnpooled().compose(sends);
-
-            assertEquals(65, comp.countComponents());
-            buffer.addMessage(comp, comp.readableBytes(), executor.newPromise());
-            assertEquals(0, buffer.nioBufferCount(), "Should still be 0 as not flushed yet");
-            buffer.addFlush();
-            final int maxCount = 10;    // less than comp.nioBufferCount()
-            ByteBuffer[] buffers = buffer.nioBuffers(maxCount, Integer.MAX_VALUE);
-            assertTrue(buffer.nioBufferCount() <= maxCount, "Should not be greater than maxCount");
-            try (var iteration = buf.forEachComponent()) {
-                var component = iteration.firstReadable();
-                ByteBuffer expected = component.readableBuffer();
-                for (int i = 0; i < buffer.nioBufferCount(); i++) {
-                    assertEquals(expected, buffers[i]);
-                }
-                assertNull(component.nextReadable());
-            }
-            buf.close();
         });
     }
